teht 4.15-4.22

+++ osa4/blogilista-back/controllers/blogs.js
@@ -1,10 +1,14 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/middleware').userExtractor
 
-blogRouter.get('/', (request, response) => {
-    Blog.find({}).then(blogs => {
-        response.json(blogs.map(blog => blog.toJSON()))
-    })
+blogRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1})
+
+  response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogRouter.get('/:id', async (request, response) => {
@@ -16,23 +20,36 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
+
+    const user = request.user
   
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
       votes: 0,
+      user: user._id,
     })
   
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.json(savedBlog.toJSON())
 })
   
-  blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+  blogRouter.delete('/:id', userExtractor, async (request, response) => {
+    const userid = request.user.id
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === userid.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({error: 'Invalid user'})
+    }
   })
   
   blogRouter.put('/:id', async (request, response) => {
